Setting Up Pygame
Numerous changes made to the script revolved around getting the visual aspect of the game set up and running.
The game state and relevant game actions are now contained within the class which will allow for both a visual game to be played and a non-visual game to be played, with the key method update_ui being the differentiator.
This state of code currently only allows for a visual game to be played due to the loop within main.py.

Please not the current graphics are only to act as a placeholder and is rudimentary.

run_game will likely be overwritten as the elements within this function is now contained within run_pygame.

Next steps will be to flesh out the remaining game loop and make it functional for the number of players within the game, not just player 1.