Basic Gameplay Loop Complete
Filled out large portion of the gameplay loop using the methods generated in previous commit. Some aspects are still missing such as the trading and mortgaging of properties as this will likely need some interface to work with.
Removed board from game_state class and replaced with simpler player_position, which tracks the location of each player in a list, listing the board position as a value. Adjusted chance and community cards accordingly
Also changed from using a player system of 1 - 4 to 0 - 3 to be more consistent with the rest of the code. One exception to this is the property ownership list as this uses 0 as an indicator of un-owned.