Jail Handling and Card Sampling
Adding checks for jail where a player is not permitted to move unless they have a get out of jail free card, roll a double, or pay the 50. As with the other user based decisions, this is currently hardcoded.
Also extended the community and chance card piles so they sample from the available cards in the deck.
Some notable changes include moving the checks for bankruptcy and checking for which player is next to methods within the game_state class as these are now repeated in several locations