All Non-User Input Programming Done
Fixed bug where the number of turns in jail was not being reset to 0 as it should have been

Chance and Community cards now have their cards shuffled at the beginning of the game, and a counter acts as a method for placing the cards to the bottom of the pile as should happen in a real game of monopoly. This is handled with properties in the game_state class, specifically, chance_draw_order and chance_card_number.
Moved all previous position information into relevant methods of the game_state class.

Some small adjustments, such as displaying the roll of a player going to or in jail
Also removed the hard coded hotels on every property to prolong games in order to test the game works as intended.